requirements.txt
yfinance
pandas 
matplotlib.pyplot 
altair 
altair_viewer
streamlit 